Total Area Normalisation

Total area normalisation (TA) scales spectra so that they all have a total integral of one

taNorm(X)

Arguments

X
A numerical matrix with rows being the spectra and columns being the chemical shift variables

Value

A list of

The normalised X matrix in the first list element, and

A numerical array of the corresponding dilution factors.

Following the example below will extract the results quickly and easily.

Details

How It Works:

The integral of each experimental spectra in X is calculated

Each intensity of a spectrum is divided by it's integral which reduces it's total area to one

Background Information:

TA was once considered the gold standard in NMR spectra normalisation.

TA is also referred to as constant sum or integral normalisation

TA is useful for spectra that have similar peaks.

Advantages:

TA is a straightforward normalisation method with limited computational requirements making it quick and uncomplicated.

Combining both TA and another normalisation method could also provide more robust results.

Limitations:

Large variation in the peaks present in a spectra severly impact TA's performance.

For instance, some spectra have large peaks created by drug metabolites, etc., which increase the total area of the spectrum.

This large total area will then dramatically down scale spectra and give misleading results.

See also

Author

kylebario1@gmail.com

Examples

ta <- taNorm(X)
Xn <- ta$Xn
dilf <- ta$dilf